requirements.txt
# Core LangChain packages
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-chroma==0.1.2
langchain-groq==0.1.9
langchain-huggingface==0.0.3

# Vector store and embeddings
chromadb==0.4.22
hnswlib==0.8.0
sentence-transformers==2.6.0

# ML/AI packages
torch>=2.0.0
transformers>=4.36.0

# Pydantic
pydantic==2.7.4
pydantic-settings==2.1.0

# Web framework
streamlit==1.29.0

# Data handling
pandas==2.1.4
numpy==1.26.3

# Utilities
python-dotenv==1.0.0